assets/logo/pyproject.toml
[project]
name = "logo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.5",
    "pillow>=12.0.0",
]
